BUG: move module to avoid circular imports
