+++ src/tike/operators/cupy/objective.py
